Create new model and relationships to help display measurements.
To make a list display, measurements should be organized by date. A new data class - MeasurementSet implements this. Time and aquarium information is stored in this class. MeasurementSet has a 1 to many relationship with Measurement, which is tracked in the MeasurementsByDate class. Finally, the data class AquariumWithMeasurements can be used to get all measurements associated with an aquarium grouped by date.

From Android documentation:
"Caution: Querying data with nested relationships requires Room to manipulate a large volume of data and can affect performance. Use as few nested relationships as possible in your queries."
However without structuring the relationships like this the measurements would need to be processed and organized at display time. If performance is too bad, maybe limiting the number of measurements can improve.