Create list of measurements by date under chart.
Renamed ParameterChart.kt to ParameterDetail.kt and added a recycler view to hold the measurements. Using the MeasurementsByDate model, each date gets a line with the measurements displayed in Parameter order. There is a title row where the color from the graph is used to mark which parameter is displayed below. The measurements and parameter colors are textviews that are generated programmatically and added to layout in the fragment.